Split the file in 2 for better future pull requests
